+++ crates/templated_uri/src/base_path.rs
